refactor: move createBundles.js to internal
