+++ packages/deploy-script-support/package.json
@@ -37,6 +37,7 @@
     "@agoric/assert": "^0.5.1",
     "@agoric/ertp": "^0.15.3",
     "@agoric/import-manager": "^0.3.6",
+    "@agoric/internal": "^0.2.1",
     "@agoric/nat": "^4.1.0",
     "@agoric/notifier": "^0.5.1",
     "@agoric/store": "^0.8.3",
