+++ packages/deploy-script-support/src/helpers.js
@@ -11,7 +11,7 @@ import { assertOfferResult } from './assertOfferResult.js';
 import { installInPieces } from './installInPieces.js';
 import { makeWriteCoreProposal } from './writeCoreProposal.js';
 
-export * from './createBundles.js';
+export * from '@agoric/internal/src/createBundles.js';
 
 // These are also hard-coded in lib-wallet.js.
 // TODO: Add methods to the wallet to access these without hard-coding
