+++ packages/governance/scripts/build-bundles.js
@@ -1,6 +1,6 @@
 #! /usr/bin/env node
 import '@endo/init';
-import { createBundles } from '@agoric/deploy-script-support';
+import { createBundles } from '@agoric/internal/src/createBundles.js';
 import url from 'url';
 
 const dirname = url.fileURLToPath(new URL('.', import.meta.url));
