+++ packages/smart-wallet/package.json
@@ -17,7 +17,6 @@
   },
   "devDependencies": {
     "@agoric/cosmic-proto": "^0.2.1",
-    "@agoric/deploy-script-support": "^0.9.4",
     "@endo/captp": "^2.0.18",
     "@endo/init": "^0.5.52",
     "ava": "^5.1.0"
